fix: add CameraPhantomInt class
